requirements.txt
absl-py==0.12.0
av==8.0.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
docopt==0.6.2
fvcore==0.1.5.post20210423
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
gulpio==540.66
idna==2.10
iopath==0.1.8
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.2.3
parameterized==0.8.1
Pillow==8.2.0
portalocker==2.3.0
protobuf==3.15.8
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorchvideo==0.1.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
sh==1.14.1
six==1.15.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.8.1
torch-tb-profiler==0.1.0
torchvideo==0.0.0
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
yacs==0.1.8
